extract git config tasks into GitConfigPlugin and make use of it in GitSetupPlugin and UpgradeDependencyPlugin
