+++ subprojects/shipkit/src/main/groovy/org/shipkit/internal/gradle/versionupgrade/UpgradeDependencyPlugin.java
@@ -10,6 +10,7 @@ import org.shipkit.gradle.exec.ShipkitExecTask;
 import org.shipkit.gradle.git.GitPushTask;
 import org.shipkit.internal.gradle.configuration.DeferredConfiguration;
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin;
+import org.shipkit.internal.gradle.git.GitConfigPlugin;
 import org.shipkit.internal.gradle.git.GitOriginPlugin;
 import org.shipkit.internal.gradle.git.GitUrlInfo;
 import org.shipkit.internal.gradle.git.tasks.GitCheckOutTask;
@@ -80,6 +81,7 @@ public class UpgradeDependencyPlugin implements Plugin<Project> {
         IncubatingWarning.warn("upgrade-dependency plugin");
         final GitOriginPlugin gitOriginPlugin = project.getRootProject().getPlugins().apply(GitOriginPlugin.class);
         final ShipkitConfiguration conf = project.getPlugins().apply(ShipkitConfigurationPlugin.class).getConfiguration();
+        project.getPlugins().apply(GitConfigPlugin.class);
 
         upgradeDependencyExtension = project.getExtensions().create("upgradeDependency", UpgradeDependencyExtension.class);
 
@@ -152,6 +154,7 @@ public class UpgradeDependencyPlugin implements Plugin<Project> {
             public void execute(final ShipkitExecTask exec) {
                 exec.setDescription("Commits updated build file.");
                 exec.mustRunAfter(REPLACE_VERSION);
+                exec.dependsOn(GitConfigPlugin.SET_EMAIL_TASK, GitConfigPlugin.SET_USER_TASK);
 
                 DeferredConfiguration.deferredConfiguration(project, new Runnable() {
                     @Override
